Refactor: Integrate activity creation into modal
