+++ src/components/modals/DealDetailsModal.tsx
@@ -16,7 +16,6 @@ import { useToast } from "@/hooks/use-toast";
 import { AddTagModal } from "./AddTagModal";
 import { AddContactTagButton } from "@/components/chat/AddContactTagButton";
 import { CreateActivityModal } from "./CreateActivityModal";
-import { ActivityForm } from "./ActivityForm";
 import { format } from "date-fns";
 import { ptBR } from "date-fns/locale";
 import { usePipelinesContext } from "@/contexts/PipelinesContext";
@@ -80,6 +79,22 @@ export function DealDetailsModal({
   const [activities, setActivities] = useState<Activity[]>([]);
   const [showAddTagModal, setShowAddTagModal] = useState(false);
   const [showCreateActivityModal, setShowCreateActivityModal] = useState(false);
+  
+  // Estados para o formulário de atividade integrado
+  const [activityForm, setActivityForm] = useState({
+    type: "Lembrete",
+    responsibleId: "",
+    subject: "",
+    description: "",
+    durationMinutes: 30,
+  });
+  const [selectedDate, setSelectedDate] = useState<Date>(new Date());
+  const [selectedTime, setSelectedTime] = useState("13:00");
+  const [users, setUsers] = useState<{ id: string; name: string; }[]>([]);
+  const [attachedFile, setAttachedFile] = useState<File | null>(null);
+  const [isCreatingActivity, setIsCreatingActivity] = useState(false);
+  const [showTimePicker, setShowTimePicker] = useState(false);
+  const [isLoadingUsers, setIsLoadingUsers] = useState(false);
   const [isLoadingData, setIsLoadingData] = useState(false);
   const [pipelineSteps, setPipelineSteps] = useState<PipelineStep[]>([]);
   const [contactPipelines, setContactPipelines] = useState<any[]>([]);
@@ -281,21 +296,27 @@ export function DealDetailsModal({
       console.error('Erro ao buscar atividades:', error);
     }
   };
-  // Função para carregar usuários apenas quando necessário
+  // Função para carregar usuários
   const fetchUsers = async () => {
+    setIsLoadingUsers(true);
     try {
       const { data, error } = await supabase
         .from('system_users')
         .select('id, name')
         .eq('status', 'active')
-        .order('name')
-        .limit(20);
+        .order('name');
         
       if (error) throw error;
-      return data || [];
+      setUsers(data?.map(user => ({ id: user.id, name: user.name })) || []);
     } catch (error) {
       console.error('Erro ao buscar usuários:', error);
-      return [];
+      toast({
+        title: "Erro",
+        description: "Não foi possível carregar os usuários.",
+        variant: "destructive",
+      });
+    } finally {
+      setIsLoadingUsers(false);
     }
   };
   const handleTagAdded = (tag: Tag) => {
@@ -323,6 +344,121 @@ export function DealDetailsModal({
   const handleActivityCreated = (activity: Activity) => {
     setActivities(prev => [...prev, activity]);
   };
+
+  // Funções para o formulário de atividade integrado
+  const handleTimeClick = () => {
+    setShowTimePicker(!showTimePicker);
+  };
+
+  const handleTimeSelect = (hour: number, minute: number) => {
+    const timeString = `${hour.toString().padStart(2, '0')}:${minute.toString().padStart(2, '0')}`;
+    setSelectedTime(timeString);
+    setShowTimePicker(false);
+  };
+
+  const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const file = event.target.files?.[0];
+    if (file) {
+      setAttachedFile(file);
+    }
+  };
+
+  const removeFile = () => {
+    setAttachedFile(null);
+  };
+
+  const handleCreateActivity = async () => {
+    if (!selectedDate || !activityForm.responsibleId || !activityForm.subject.trim()) {
+      toast({
+        title: "Campos obrigatórios",
+        description: "Preencha todos os campos obrigatórios.",
+        variant: "destructive",
+      });
+      return;
+    }
+
+    setIsCreatingActivity(true);
+    try {
+      // Combinar data e hora
+      const [hour, minute] = selectedTime.split(':').map(Number);
+      const scheduledDateTime = new Date(selectedDate);
+      scheduledDateTime.setHours(hour, minute, 0, 0);
+
+      // Get workspace_id from the contact
+      const { data: contactDataForActivity } = await supabase
+        .from('contacts')
+        .select('workspace_id')
+        .eq('id', contactId)
+        .single();
+
+      if (!contactDataForActivity?.workspace_id) {
+        throw new Error('Workspace não encontrado para este contato');
+      }
+
+      const activityData = {
+        contact_id: contactId,
+        workspace_id: contactDataForActivity.workspace_id,
+        type: activityForm.type,
+        responsible_id: activityForm.responsibleId,
+        subject: activityForm.subject,
+        description: activityForm.description || null,
+        scheduled_for: scheduledDateTime.toISOString(),
+        duration_minutes: activityForm.durationMinutes,
+        attachment_name: attachedFile?.name || null,
+        attachment_url: null,
+      };
+
+      const { data: activity, error } = await supabase
+        .from('activities')
+        .insert(activityData)
+        .select(`
+          id,
+          type,
+          subject,
+          scheduled_for,
+          responsible_id,
+          is_completed
+        `)
+        .single();
+
+      if (error) throw error;
+
+      handleActivityCreated(activity);
+      
+      toast({
+        title: "Atividade criada com sucesso!",
+        description: `A atividade "${activityForm.subject}" foi agendada.`,
+      });
+
+      // Resetar formulário
+      setActivityForm({
+        type: "Lembrete",
+        responsibleId: "",
+        subject: "",
+        description: "",
+        durationMinutes: 30,
+      });
+      setSelectedDate(new Date());
+      setSelectedTime("13:00");
+      setAttachedFile(null);
+    } catch (error) {
+      console.error('Erro ao criar atividade:', error);
+      toast({
+        title: "Erro",
+        description: "Não foi possível criar a atividade.",
+        variant: "destructive",
+      });
+    } finally {
+      setIsCreatingActivity(false);
+    }
+  };
+
+  // Carregar usuários quando necessário
+  useEffect(() => {
+    if (activeTab === "atividades" && users.length === 0) {
+      fetchUsers();
+    }
+  }, [activeTab]);
   const handleCompleteActivity = async (activityId: string) => {
     try {
       const {
@@ -630,12 +766,239 @@ export function DealDetailsModal({
                   </div>}
               </div>
 
-              {/* Formulário Criar Atividade */}
-              <ActivityForm 
-                contactId={contactId} 
-                isDarkMode={isDarkMode} 
-                onActivityCreated={handleActivityCreated}
-              />
+              {/* Formulário Criar Atividade - Integrado */}
+              <div className="space-y-4">
+                <h3 className={cn("text-lg font-semibold mb-4", isDarkMode ? "text-white" : "text-gray-900")}>
+                  Criar atividade
+                </h3>
+                
+                <div className="space-y-4">
+                  {/* Tipo */}
+                  <div className="space-y-2">
+                    <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                      Tipo
+                    </label>
+                    <Select value={activityForm.type} onValueChange={(value) => setActivityForm({...activityForm, type: value})}>
+                      <SelectTrigger className={cn("w-full", isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white" : "bg-white")}>
+                        <div className="flex items-center gap-2">
+                          {(() => {
+                            const activityTypes = [
+                              { value: "Lembrete", label: "Lembrete", icon: Clock },
+                              { value: "Mensagem", label: "Mensagem", icon: MessageSquare },
+                              { value: "Ligação", label: "Ligação", icon: Phone },
+                              { value: "Reunião", label: "Reunião", icon: User },
+                              { value: "Agendamento", label: "Agendamento", icon: CalendarIcon },
+                            ];
+                            const selectedType = activityTypes.find(t => t.value === activityForm.type);
+                            const Icon = selectedType?.icon || Clock;
+                            return (
+                              <>
+                                <Icon className="w-4 h-4" />
+                                <span>{selectedType?.label || activityForm.type}</span>
+                              </>
+                            );
+                          })()}
+                        </div>
+                      </SelectTrigger>
+                      <SelectContent>
+                        {[
+                          { value: "Lembrete", label: "Lembrete", icon: Clock },
+                          { value: "Mensagem", label: "Mensagem", icon: MessageSquare },
+                          { value: "Ligação", label: "Ligação", icon: Phone },
+                          { value: "Reunião", label: "Reunião", icon: User },
+                          { value: "Agendamento", label: "Agendamento", icon: CalendarIcon },
+                        ].map((type) => {
+                          const Icon = type.icon;
+                          return (
+                            <SelectItem key={type.value} value={type.value}>
+                              <div className="flex items-center gap-2">
+                                <Icon className="w-4 h-4" />
+                                <span>{type.label}</span>
+                              </div>
+                            </SelectItem>
+                          );
+                        })}
+                      </SelectContent>
+                    </Select>
+                  </div>
+
+                  {/* Responsável */}
+                  <div className="space-y-2">
+                    <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                      Responsável
+                    </label>
+                    <Select value={activityForm.responsibleId} onValueChange={(value) => setActivityForm({...activityForm, responsibleId: value})}>
+                      <SelectTrigger className={cn("w-full", isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white" : "bg-white")}>
+                        <SelectValue placeholder={isLoadingUsers ? "Carregando usuários..." : "Selecione um responsável"} />
+                      </SelectTrigger>
+                      <SelectContent>
+                        {users.map((user) => (
+                          <SelectItem key={user.id} value={user.id}>
+                            {user.name}
+                          </SelectItem>
+                        ))}
+                      </SelectContent>
+                    </Select>
+                  </div>
+
+                  {/* Assunto */}
+                  <div className="space-y-2">
+                    <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                      Assunto
+                    </label>
+                    <Input 
+                      placeholder="Digite o assunto da atividade" 
+                      value={activityForm.subject}
+                      onChange={(e) => setActivityForm({...activityForm, subject: e.target.value})}
+                      className={cn(isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white" : "bg-white")} 
+                    />
+                  </div>
+
+                  {/* Data e Duração em linha */}
+                  <div className="grid grid-cols-2 gap-3">
+                    <div className="space-y-2">
+                      <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                        Agendar para
+                      </label>
+                      <Popover>
+                        <PopoverTrigger asChild>
+                          <Button variant="outline" className={cn("w-full justify-start text-left font-normal", isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white hover:bg-gray-700" : "bg-white")}>
+                            <CalendarIcon className="mr-2 h-4 w-4" />
+                            {selectedDate ? format(selectedDate, "dd/MM/yyyy", { locale: ptBR }) : "Selecionar data"}
+                          </Button>
+                        </PopoverTrigger>
+                        <PopoverContent className="w-auto p-0" align="start">
+                          <Calendar 
+                            mode="single" 
+                            selected={selectedDate} 
+                            onSelect={(date) => date && setSelectedDate(date)} 
+                            initialFocus 
+                            className="pointer-events-auto" 
+                          />
+                        </PopoverContent>
+                      </Popover>
+                    </div>
+
+                    <div className="space-y-2">
+                      <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                        Duração (minutos)
+                      </label>
+                      <Input 
+                        type="number" 
+                        value={activityForm.durationMinutes} 
+                        onChange={(e) => setActivityForm({...activityForm, durationMinutes: Number(e.target.value)})}
+                        className={cn(isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white" : "bg-white")} 
+                      />
+                    </div>
+                  </div>
+
+                  {/* Hora */}
+                  <div className="space-y-2">
+                    <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                      Hora
+                    </label>
+                    <Popover open={showTimePicker} onOpenChange={setShowTimePicker}>
+                      <PopoverTrigger asChild>
+                        <Button
+                          variant="outline"
+                          className={cn(
+                            "w-full justify-start text-left font-normal",
+                            isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white" : "bg-white"
+                          )}
+                          onClick={handleTimeClick}
+                        >
+                          <Clock className="mr-2 h-4 w-4" />
+                          {selectedTime}
+                        </Button>
+                      </PopoverTrigger>
+                      <PopoverContent className="w-auto p-0" align="start">
+                        <div className="max-h-60 overflow-y-auto p-2">
+                          {(() => {
+                            const timeOptions = [];
+                            for (let hour = 0; hour < 24; hour++) {
+                              for (let minute = 0; minute < 60; minute += 30) {
+                                timeOptions.push({
+                                  hour,
+                                  minute,
+                                  display: `${hour.toString().padStart(2, '0')}:${minute.toString().padStart(2, '0')}`
+                                });
+                              }
+                            }
+                            return timeOptions.map((time) => (
+                              <Button
+                                key={time.display}
+                                variant="ghost"
+                                size="sm"
+                                className="w-full justify-start"
+                                onClick={() => handleTimeSelect(time.hour, time.minute)}
+                              >
+                                {time.display}
+                              </Button>
+                            ));
+                          })()}
+                        </div>
+                      </PopoverContent>
+                    </Popover>
+                  </div>
+
+                  {/* Upload de arquivo */}
+                  <div className="space-y-2">
+                    <div className={cn("border-2 border-dashed rounded-lg p-6 text-center cursor-pointer transition-colors", isDarkMode ? "border-gray-600 hover:border-gray-500 bg-[#1f1f1f]" : "border-gray-300 hover:border-gray-400 bg-gray-50")}>
+                      {attachedFile ? (
+                        <div className="flex items-center justify-between">
+                          <span className={cn("text-sm", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                            {attachedFile.name}
+                          </span>
+                          <Button
+                            size="sm"
+                            variant="ghost"
+                            onClick={removeFile}
+                            className="h-6 w-6 p-0"
+                          >
+                            <X className="w-4 h-4" />
+                          </Button>
+                        </div>
+                      ) : (
+                        <label className="cursor-pointer">
+                          <input
+                            type="file"
+                            onChange={handleFileUpload}
+                            className="hidden"
+                            accept="*/*"
+                          />
+                          <Upload className={cn("w-8 h-8 mx-auto mb-2", isDarkMode ? "text-gray-400" : "text-gray-500")} />
+                          <p className={cn("text-sm", isDarkMode ? "text-gray-400" : "text-gray-600")}>
+                            Clique aqui ou arraste o documento a ser salvo
+                          </p>
+                        </label>
+                      )}
+                    </div>
+                  </div>
+
+                  {/* Descrição */}
+                  <div className="space-y-2">
+                    <label className={cn("text-sm font-medium", isDarkMode ? "text-gray-300" : "text-gray-700")}>
+                      Descrição
+                    </label>
+                    <Textarea 
+                      placeholder="Descrição" 
+                      rows={4} 
+                      value={activityForm.description}
+                      onChange={(e) => setActivityForm({...activityForm, description: e.target.value})}
+                      className={cn(isDarkMode ? "bg-[#2d2d2d] border-gray-600 text-white" : "bg-white")} 
+                    />
+                  </div>
+
+                  {/* Botão Criar Atividade */}
+                  <Button 
+                    className="w-full bg-blue-600 hover:bg-blue-700 text-white" 
+                    onClick={handleCreateActivity} 
+                    disabled={!contactId || isCreatingActivity}
+                  >
+                    {isCreatingActivity ? "Criando..." : "Criar"}
+                  </Button>
+                </div>
+              </div>
             </div>}
 
           {activeTab === "historico" && <div className="space-y-4">
